REFACTOR-#0000: move lazy_metadata_decorator into utils.py
